Refactor #7 - Move results to it's own file
